+++ executors.js
@@ -12,7 +12,7 @@ var __importDefault = (this && this.__importDefault) || function (mod) {
     return (mod && mod.__esModule) ? mod : { "default": mod };
 };
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.do_post = exports.do_help = exports.do_queue_time = exports.printTableHeader = exports.alignLeft = exports.alignRight = exports.do_cron = exports.do_event = exports.do_envvar = exports.do_key = exports.do_inspect = exports.do_build = exports.do_redeploy = exports.do_deploy = exports.generateNewKey = exports.useExistingKey = exports.do_register = exports.addKnownHost = exports.do_duplicate = exports.fetch_template = exports.createService = exports.createServiceGroup = exports.createOrganization = exports.do_clone = exports.do_fetch = void 0;
+exports.do_remove_auto_approve = exports.do_auto_approve = exports.do_attach_role = exports.do_join = exports.do_post = exports.do_help = exports.do_queue_time = exports.printTableHeader = exports.alignLeft = exports.alignRight = exports.do_cron = exports.do_event = exports.do_envvar = exports.do_key = exports.do_inspect = exports.do_build = exports.do_redeploy = exports.do_deploy = exports.generateNewKey = exports.useExistingKey = exports.do_register = exports.addKnownHost = exports.do_duplicate = exports.fetch_template = exports.createService = exports.createServiceGroup = exports.createOrganization = exports.do_clone = exports.do_fetch = void 0;
 const fs_1 = __importDefault(require("fs"));
 const os_1 = __importDefault(require("os"));
 const utils_1 = require("./utils");
@@ -603,3 +603,47 @@ function do_post(eventType, key, contentType, payload) {
     });
 }
 exports.do_post = do_post;
+function do_join(org) {
+    return __awaiter(this, void 0, void 0, function* () {
+        try {
+            (0, utils_1.output2)(yield (0, utils_1.sshReq)(`org`, `--join`, org));
+        }
+        catch (e) {
+            throw e;
+        }
+    });
+}
+exports.do_join = do_join;
+function do_attach_role(org, user, role) {
+    return __awaiter(this, void 0, void 0, function* () {
+        try {
+            (0, utils_1.output2)(yield (0, utils_1.sshReq)(`role`, `--user`, user, `--org`, org, role));
+        }
+        catch (e) {
+            throw e;
+        }
+    });
+}
+exports.do_attach_role = do_attach_role;
+function do_auto_approve(org, domain, role) {
+    return __awaiter(this, void 0, void 0, function* () {
+        try {
+            (0, utils_1.output2)(yield (0, utils_1.sshReq)(`preapprove`, `--add`, role, `--org`, org, domain));
+        }
+        catch (e) {
+            throw e;
+        }
+    });
+}
+exports.do_auto_approve = do_auto_approve;
+function do_remove_auto_approve(org, domain) {
+    return __awaiter(this, void 0, void 0, function* () {
+        try {
+            (0, utils_1.output2)(yield (0, utils_1.sshReq)(`preapprove`, `--org`, org, domain));
+        }
+        catch (e) {
+            throw e;
+        }
+    });
+}
+exports.do_remove_auto_approve = do_remove_auto_approve;